move likeBlockStoreTest module to test folder, import the module if needed
