+++ movie_list/src/App.js
@@ -6,11 +6,7 @@ import MovieList from './components/movieListComponent/layout';
 import MovieModal from './components/movie_modal';
 import Sidebar from './components/sidebar/sidebar';
 
-import configureStore from './store/store';
-import { likeMovie, unlikeMovie, blockMovie, unblockMovie } from './actionCreators/like_block_action_creator';
-const myStore = configureStore(); //Assume the buttons are connected to temporary redux store, need to connect the component to real store later
-let prevLikedList = {};
-let prevBlockedList = {}; //check reference
+import LikeBlockStoreTest from './tests/likeBlockStoreTest';
 function App() {
   const [showOption,setShow] = useState({show:false,movieID:null});
   return (
@@ -18,54 +14,7 @@ function App() {
       <Router>
         <Sidebar/>
       </Router>
-      <button onClick={()=>{
-        prevLikedList = myStore.getState().likeBlockLists.likedList;
-        prevBlockedList = myStore.getState().likeBlockLists.blockedList;
-        myStore.dispatch(likeMovie(278,1));/*get movieID and pageNum from props*/
-        console.log("prevLikedList: ",prevLikedList);
-        console.log("prevBlockedList: ",prevBlockedList);
-        console.log("likedlist now: ",myStore.getState().likeBlockLists.likedList);
-        console.log("blockedlist now: ",myStore.getState().likeBlockLists.blockedList);
-        console.log("Same Liked List: ",prevLikedList === myStore.getState().likeBlockLists.likedList);
-        console.log("Same Blocked List: ",prevBlockedList === myStore.getState().likeBlockLists.blockedList);}}>
-        like
-      </button>
-      <button onClick={()=>{
-        prevLikedList = myStore.getState().likeBlockLists.likedList;
-        prevBlockedList = myStore.getState().likeBlockLists.blockedList;
-        myStore.dispatch(unlikeMovie(278,1));
-        console.log("prevLikedList: ",prevLikedList);
-        console.log("prevBlockedList: ",prevBlockedList);
-        console.log("likedlist now: ",myStore.getState().likeBlockLists.likedList);
-        console.log("blockedlist now: ",myStore.getState().likeBlockLists.blockedList);
-        console.log("Same Liked List: ",prevLikedList === myStore.getState().likeBlockLists.likedList);
-        console.log("Same Blocked List: ",prevBlockedList === myStore.getState().likeBlockLists.blockedList);}}>
-        unlike
-      </button>
-      <button onClick={()=>{
-        prevLikedList = myStore.getState().likeBlockLists.likedList;
-        prevBlockedList = myStore.getState().likeBlockLists.blockedList;
-        myStore.dispatch(blockMovie(278,1));
-        console.log("prevLikedList: ",prevLikedList);
-        console.log("prevBlockedList: ",prevBlockedList);
-        console.log("likedlist now: ",myStore.getState().likeBlockLists.likedList);
-        console.log("blockedlist now: ",myStore.getState().likeBlockLists.blockedList);
-        console.log("Same Liked List: ",prevLikedList === myStore.getState().likeBlockLists.likedList);
-        console.log("Same Blocked List: ",prevBlockedList === myStore.getState().likeBlockLists.blockedList);}}>
-        block
-      </button>
-      <button onClick={()=>{
-        prevLikedList = myStore.getState().likeBlockLists.likedList;
-        prevBlockedList = myStore.getState().likeBlockLists.blockedList;
-        myStore.dispatch(unblockMovie(278,1));
-        console.log("prevLikedList: ",prevLikedList);
-        console.log("prevBlockedList: ",prevBlockedList);
-        console.log("likedlist now: ",myStore.getState().likeBlockLists.likedList);
-        console.log("blockedlist now: ",myStore.getState().likeBlockLists.blockedList);
-        console.log("Same Liked List: ",prevLikedList === myStore.getState().likeBlockLists.likedList);
-        console.log("Same Blocked List: ",prevBlockedList === myStore.getState().likeBlockLists.blockedList);}}>
-        unblock
-      </button>
+      <LikeBlockStoreTest/>
       <MovieList/>
       <MovieModal showOption={showOption} setShow={setShow}/>
     </div>
